fix: CPL double $ + Int64 tokens + multi-range snapshot merge + macOS no-stale flash
- iOS CPL row: removed extra $ before usd() (was "CPL $$0.31")
- DashboardSnapshot: todayCalls/todayTokens → Int64 (arm64_32 overflow)
- TrendPoint: calls/tokens → Int64
- iOS checkCloud(): fetch week + 30d snapshots after hasData()
- watchOS refresh(): fetchAndMerge today→week→30d with range-specific fields
- macOS: load() then pollAll(), skip cache on initial load

+++ Suites/Shared/Models/DashboardSnapshot.swift
@@ -11,8 +11,8 @@ struct DashboardSnapshot: Codable {
     var yesterdaySpend: Double = 0
     var previousPeriodSpend: Double = 0
     var subDaily: Double = 0
-    var todayCalls: Int = 0
-    var todayTokens: Int = 0
+    var todayCalls: Int64 = 0
+    var todayTokens: Int64 = 0
 
     var providerBreakdown: [ProviderItem] = []
     var toolBreakdown: [NameCostItem] = []
@@ -55,8 +55,8 @@ struct PredictionItem: Codable {
 struct TrendPoint: Codable {
     var ts: Double
     var value: Double
-    var calls: Int
-    var tokens: Int
+    var calls: Int64
+    var tokens: Int64
     var netLines: Int
     var added: Int = 0
     var deleted: Int = 0
